// frontend/src/model/groundStationEntity.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * My Watch Tower API
 * Satellite tracker and predictor
 * OpenAPI spec version: 1.0
 */

export interface GroundStationEntity {
  id: number;
  name: string;
  latitude: number;
  longitude: number;
  altitude: number;
  horizonmask: string;
  isDefault: boolean;
  createdAt: string;
  updatedAt: string;
}